# server/server.js
require('./config/config');
require('./db/mongoose');

const _ = require('lodash');
const express = require('express');
const bodyParser = require('body-parser');
const { ObjectID } = require('mongodb');


const swaggerUi = require('swagger-ui-express');
const swaggerDocument = require('./../APIDefinition.json');


const { Todo } = require('./models/todo');
const { User } = require('./models/user');

const { authenticate } = require('./middlewares/authenticate');


let app = express();
app.disable('x-powered-by');
const port = process.env.PORT;

app.use(bodyParser.json());
app.use('/', swaggerUi.serve, swaggerUi.setup(swaggerDocument));


app.post('/todos', authenticate, (req, res) => {
    let todo = new Todo({
        text: req.body.text,
        _creator: req.user._id
    });

    todo.save().then(doc => {
        res.send(doc);
    }, (e) => {
        res.status(400).send(e);
    });
});


app.get('/todos', authenticate, (req, res) => {
    Todo.find({
        _creator: req.user._id
    }).then(todos => {
        res.send({ todos })
    }, (e) => {
        res.status(400).send(e);
    });
});


app.get('/todos/:id', authenticate, (req, res) => {
    let id = req.params.id;

    if(!ObjectID.isValid(id))
        return res.status(404).send();

    Todo.findOne({
        _id: id,
        _creator: req.user._id
    }).then((todo) => {
        if(todo)
            return res.send({ todo });
        res.status(404).send();
    }).catch(() => {
        res.status(400).send();
    });
});


app.delete('/todos/:id', authenticate, (req, res) => {
    let id = req.params.id;

    if(!ObjectID.isValid(id))
        return res.status(404).send();

    Todo.findOneAndRemove({
        _id: id,
        _creator: req.user._id
    }).then((todo) => {
        if(todo)
            return res.send({ todo });
        res.status(404).send();
    }).catch(() => {
        res.status(400).send();
    });
});


app.patch('/todos/:id', authenticate, (req, res) => {
    let id = req.params.id;
    let body = _.pick(req.body, ['text', 'completed']);
    
    if(!ObjectID.isValid(id))
        return res.status(404).send();

    if(_.isBoolean(body.completed) && body.completed) {
        body.completedAt = new Date().getTime();
    } else {
        body.completed = false;
        body.completedAt = null;
    }

    Todo.findOneAndUpdate({
        _id: id,
        _creator: req.user._id
    }, {$set: body}, {new: true}).then((todo) => {
        if(!todo)
            return res.status(404).send();
        res.send({ todo });
    }).catch((e) => {
        res.status(400).send();
    });
});


// POST /users
app.post('/users', (req, res) => {
    let user = new User({
        email: req.body.email,
        password: req.body.password
    });

    user.save().then(() => {
        return user.generateAuthToken();
    })
    .then(token => {
        res.header('x-auth', token).send(user);
    })
    .catch(e => {
        res.status(400).send(e);
    });
});


app.get('/users/me', authenticate, (req, res) => {
   res.send(req.user);
});


// POST /users/login {email, password}
app.post('/users/login', (req, res) => {
    let { email, password } = req.body;
    
    User.findByCredentials(email, password).then(user => {
        user.generateAuthToken().then(token => {
            res.header('x-auth', token).send(user);
        });
    }).catch(e => {
        res.status(400).send();
    });
});


app.delete('/users/me/token', authenticate, (req, res) => {
    req.user.removeToken(req.token).then(() => {
        res.status(200).send();
    }, () => {
        res.status(400).send();
    });
});


app.listen(port, () => {
    console.log(`Started up at port ${port}`);
});

module.exports = { app };